refactor(e2e): focus sync.rs and take a bare fixture name in setup_project
Two small test-organization cleanups:

- Split the two fixture-based tests out of sync.rs into their own focused
  files (tests/nested.rs, tests/multi_dir.rs). sync.rs now holds only the
  static-site sync lifecycle (deploy / no-op / update / delete / proxy).
- `setup_project` now takes a bare fixture name (e.g. `setup_project("nested")`)
  and prepends `tests/fixture/` itself, mirroring `setup_example("<name>")`,
  instead of requiring the `tests/fixture/` path prefix at every call site.
