+++ crates/e2e/tests/recipe.rs
@@ -25,7 +25,7 @@ fn setup_recipe_project(fixture: &str) -> TempDir {
 /// `dir` only: deploy through the recipe and confirm the asset is served.
 #[test]
 fn recipe_basic() {
-    let tmp = setup_recipe_project("tests/fixture/recipe-basic");
+    let tmp = setup_recipe_project("recipe-basic");
     let project = tmp.path();
     let _network = LocalNetwork::start(project);
 
@@ -41,7 +41,7 @@ fn recipe_basic() {
 /// `build`: a pre-sync command produces an extra asset, which must then sync.
 #[test]
 fn recipe_build_step() {
-    let tmp = setup_recipe_project("tests/fixture/recipe-build");
+    let tmp = setup_recipe_project("recipe-build");
     let project = tmp.path();
     let _network = LocalNetwork::start(project);
 
@@ -61,7 +61,7 @@ fn recipe_build_step() {
 /// serves assets proves the injected wasm installs and runs.
 #[test]
 fn recipe_metadata() {
-    let tmp = setup_recipe_project("tests/fixture/recipe-metadata");
+    let tmp = setup_recipe_project("recipe-metadata");
     let project = tmp.path();
     let _network = LocalNetwork::start(project);
