+++ crates/e2e/tests/sync.rs
@@ -1,7 +1,5 @@
 use candid::Principal;
-use e2e::{
-    icp_cmd, list_assets, setup_example, setup_project, AssetDetails, Encoding, LocalNetwork,
-};
+use e2e::{icp_cmd, list_assets, setup_example, AssetDetails, Encoding, LocalNetwork};
 use std::fs;
 
 /// Deploy the `static-site` example to a local replica and verify that
@@ -19,30 +17,6 @@ fn basic_deploy() {
     assert!(
         assets.iter().any(|a| a.key == "/index.html"),
         "expected /index.html in canister asset list; got: {assets:#?}",
-    );
-}
-
-/// Deploy a fixture whose `dirs` entry is a *nested* path (`src/frontend/dist`).
-/// The host preopens it under a multi-segment WASI guest name; the plugin's scan
-/// step must not call `canonicalize`/`realpath` on it (WASI returns ENOENT for
-/// any path under a multi-component preopen, even though plain access works).
-#[test]
-fn nested_dir_deploy() {
-    let tmp = setup_project("tests/fixture/nested");
-    let project = tmp.path();
-    let _network = LocalNetwork::start(project);
-
-    icp_cmd(project).arg("deploy").assert().success();
-
-    let assets = list_assets(project);
-
-    assert!(
-        assets.iter().any(|a| a.key == "/index.html"),
-        "expected /index.html in canister asset list; got: {assets:#?}",
-    );
-    assert!(
-        assets.iter().any(|a| a.key == "/assets/style.css"),
-        "expected /assets/style.css (nested subdir) in canister asset list; got: {assets:#?}",
     );
 }
 
@@ -206,31 +180,5 @@ fn asset_deletion() {
     assert!(
         !assets_after.iter().any(|a| a.key == "/style.css"),
         "/style.css should be removed from the canister after local deletion",
-    );
-}
-
-/// The assets sync plugin owns the URL space of its canister and only
-/// supports a single source directory. A manifest that lists multiple
-/// `dirs:` entries must fail the sync step before any canister mutation.
-#[test]
-fn multi_directory_sync_rejected() {
-    let tmp = setup_project("tests/fixture/multi-dir");
-    let project = tmp.path();
-    let _network = LocalNetwork::start(project);
-
-    let output = icp_cmd(project)
-        .arg("deploy")
-        .assert()
-        .failure()
-        .get_output()
-        .clone();
-    let combined = format!(
-        "{}\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr),
-    );
-    assert!(
-        combined.contains("expected exactly one input directory"),
-        "expected multi-dir rejection message; got:\n{combined}",
     );
 }
